Refactor marketplace header components
